src/txts_unit/requirements.txt
click==8.1.3
hazelcast-python-client==5.1
importlib-metadata==4.11.3
mininet==2.3.0.dev6
scapy==2.4.0
typer==0.4.1
typing_extensions==4.2.0
zipp==3.8.0
Twisted==22.10.0
